week_11/CodeClanTowers/src/main/java/Bedroom.java
import java.util.ArrayList;

public class Bedroom {


    //    Bedroom which have a room number, capacity, collection of guests and a type (e.g. Single/Double.)
//
    private int roomnumber;
    private int capacity;
    private ArrayList<Guest> guests;
    private String type;

    public Bedroom(int roomnumber, int capacity, String type) {
        this.roomnumber = roomnumber;
        this.capacity = capacity;
        this.type = type;
        this.guests = new ArrayList<Guest>();


    }

    public int getRoomnumber() {
        return roomnumber;
    }

    public void setRoomnumber(int roomnumber) {
        this.roomnumber = roomnumber;
    }

    public int getCapacity() {
        return capacity;
    }

    public void setCapacity(int capacity) {
        this.capacity = capacity;
    }

    public ArrayList<Guest> getGuests() {
        return guests;
    }

    public void setGuests(ArrayList<Guest> guests) {
        this.guests = guests;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public int countGuests() {
        return this.guests.size();
    }

    public void addGuest(Guest guest) {
        this.guests.add(guest);
    }

    public void removeGuest(Guest guest) {
        this.guests.remove(guest);
    }
}
